+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * remove Color.Accent resource need
 * add chromecast as NUHF SDK
 * add ActiveJammer shadow method
+* cleaning up Active jammer code mess
 
 ### 4.4.2 (2020-8-26)
 * upgrade build target to API 29, Google Play comply